Video is working

+++ src/battle_anim_mons.c
@@ -382,7 +382,7 @@ void StoreSpriteCallbackInData6(struct Sprite *sprite, SpriteCallback callback)
 
 static void SetCallbackToStoredInData6(struct Sprite *sprite)
 {
-    u32 callback = (u16)sprite->data[6] | (sprite->data[7] << 16);
+    u32 callback = (u16)sprite->data[6] | ((u16)sprite->data[7] << 16);
     
     sprite->callback = (SpriteCallback)callback;
 }
